fix: mobile layout outside viewport
