+++ apps/evm/ui/pool/NewPositionSection/SelectNetworkWidget.tsx
@@ -1,7 +1,7 @@
 import { ChainId, chainName } from '@sushiswap/chain'
+import { NetworkSelector } from '@sushiswap/ui'
 import { Button } from '@sushiswap/ui/components/button'
 import { NetworkIcon } from '@sushiswap/ui/components/icons'
-import { NetworkSelector } from '@sushiswap/ui/components/networkselector'
 import { SelectIcon } from '@sushiswap/ui/components/select'
 import { SUSHISWAP_V3_SUPPORTED_CHAIN_IDS } from '@sushiswap/v3-sdk'
 import React, { FC, memo } from 'react'
@@ -32,7 +32,6 @@ export const SelectNetworkWidget: FC<SelectNetworkWidgetProps> = memo(function S
           networks={networks ?? SUSHISWAP_V3_SUPPORTED_CHAIN_IDS}
           selected={selectedNetwork}
           onSelect={onSelect}
-          variant="menu"
         >
           <Button variant="secondary" className="!font-medium">
             <NetworkIcon chainId={selectedNetwork} width={16} height={16} />
