+++ apps/evm/ui/swap/widget/ChainSelectors.tsx
@@ -2,9 +2,9 @@ import { ArrowRightIcon } from '@heroicons/react/24/outline'
 import { Chain } from '@sushiswap/chain'
 import { STARGATE_SUPPORTED_CHAIN_IDS, StargateChainId } from '@sushiswap/stargate'
 import { classNames } from '@sushiswap/ui'
+import { NetworkSelector, NetworkSelectorOnSelectCallback } from '@sushiswap/ui'
 import { Button } from '@sushiswap/ui/components/button'
 import { NetworkIcon } from '@sushiswap/ui/components/icons'
-import { NetworkSelector, NetworkSelectorOnSelectCallback } from '@sushiswap/ui/components/networkselector'
 import { SelectIcon } from '@sushiswap/ui/components/select'
 import React, { FC, useCallback } from 'react'
 
@@ -51,7 +51,6 @@ export const ChainSelectors: FC<{ open: boolean }> = ({ open }) => {
             <span className="text-sm">From network</span>
             <NetworkSelector
               networks={STARGATE_SUPPORTED_CHAIN_IDS}
-              variant="menu"
               selected={network0 as StargateChainId}
               onSelect={handleSelect0}
             >
@@ -66,7 +65,6 @@ export const ChainSelectors: FC<{ open: boolean }> = ({ open }) => {
             <span className="text-sm">To network</span>
             <NetworkSelector
               networks={STARGATE_SUPPORTED_CHAIN_IDS}
-              variant="menu"
               selected={network1 as StargateChainId}
               onSelect={handleSelect1}
             >
