+++ packages/ui/src/components/index.ts
@@ -30,7 +30,7 @@ export * from './loader'
 export * from './modal'
 export * from './navigation'
 export * from './navigation-menu'
-export * from './networkselector'
+export * from './network-selector'
 export * from './onramper'
 export * from './overlay'
 export * from './Popover'
