+++ packages/wagmi/src/future/components/HeaderNetworkSelector.tsx
@@ -1,11 +1,10 @@
 import { Chain, ChainId } from '@sushiswap/chain'
+import { NetworkSelector, NetworkSelectorOnSelectCallback } from '@sushiswap/ui'
+import { Button } from '@sushiswap/ui/components/button'
 import { NetworkIcon } from '@sushiswap/ui/components/icons'
-import { NetworkSelector, NetworkSelectorOnSelectCallback } from '@sushiswap/ui/components/networkselector'
 import { createErrorToast } from '@sushiswap/ui/components/toast'
-import { useBreakpoint } from '@sushiswap/ui/lib/useBreakpoint'
 import React, { FC, useCallback } from 'react'
 import { ProviderRpcError, useNetwork, UserRejectedRequestError, useSwitchNetwork } from 'wagmi'
-import { Button } from '@sushiswap/ui/components/button'
 
 export const HeaderNetworkSelector: FC<{
   networks: ChainId[]
@@ -14,7 +13,6 @@ export const HeaderNetworkSelector: FC<{
 }> = ({ networks, selectedNetwork, onChange }) => {
   const { switchNetworkAsync } = useSwitchNetwork()
   const { chain } = useNetwork()
-  const { isSm } = useBreakpoint('sm')
 
   const onSwitchNetwork = useCallback<NetworkSelectorOnSelectCallback>(
     async (el, close) => {
@@ -42,12 +40,7 @@ export const HeaderNetworkSelector: FC<{
   const selected = selectedNetwork || (chain?.id as ChainId) || ChainId.ETHEREUM
 
   return (
-    <NetworkSelector
-      selected={selected}
-      variant={isSm ? 'menu' : 'dialog'}
-      onSelect={onSwitchNetwork}
-      networks={networks}
-    >
+    <NetworkSelector selected={selected} onSelect={onSwitchNetwork} networks={networks}>
       <Button variant="secondary" testId="network-selector">
         <NetworkIcon chainId={selected} width={20} height={20} />
         <div className="hidden xl:block">{Chain.from(selected)?.name}</div>
